+++ buildpacks/ruby/src/steps.rs
@@ -1,12 +1,8 @@
-mod bundle_install;
-mod bundler_download;
 mod default_env;
 mod detect_rake_tasks;
 mod get_default_process;
 mod rake_assets_install;
 
-pub(crate) use self::bundle_install::bundle_install;
-pub(crate) use self::bundler_download::bundler_download;
 pub(crate) use self::default_env::default_env;
 pub(crate) use self::detect_rake_tasks::detect_rake_tasks;
 pub(crate) use self::get_default_process::get_default_process;