+++ commons/src/lib.rs
@@ -6,7 +6,6 @@ pub mod display;
 pub mod env_command;
 pub mod gem_list;
 pub mod gem_version;
-pub mod gemfile_lock;
 pub mod layer;
 pub mod metadata_digest;
 pub mod rake_status;